+++ contracts/fuzz/scenarios/DiffTest.sol
@@ -255,7 +255,7 @@ contract DiffTestScenario {
         uint8 tokenID,
         uint8 targetNameID,
         uint256 defaultThresholdSeed,
-        uint256 delayUntilDefaultSeed,
+        uint48 delayUntilDefaultSeed,
         bool isStable,
         bool isColl
     ) public {
@@ -291,7 +291,7 @@ contract DiffTestScenario {
         uint8 tokenID,
         uint8 targetNameID,
         uint256 defaultThresholdSeed,
-        uint256 delayUntilDefaultSeed,
+        uint48 delayUntilDefaultSeed,
         bool isStable,
         bool isColl
     ) public {
@@ -945,7 +945,7 @@ contract DiffTestScenario {
         IERC20 erc20,
         bool isStable,
         uint256 defaultThresholdSeed,
-        uint256 delayUntilDefaultSeed,
+        uint48 delayUntilDefaultSeed,
         bytes32 targetName
     ) internal returns (CollateralNoDecay) {
         lastCreatedColl = new CollateralNoDecay({
@@ -954,7 +954,7 @@ contract DiffTestScenario {
             priceTimeout_: 806400,
             oracleError_: 0.005e18,
             defaultThreshold_: uint192(between(1, 1e18, defaultThresholdSeed)),
-            delayUntilDefault_: between(1, type(uint48).max / 2, delayUntilDefaultSeed),
+            delayUntilDefault_: uint48(between(1, type(uint48).max / 2, delayUntilDefaultSeed)),
             targetName_: targetName,
             refPerTokModel_: isStable ? growing : getNextPriceModel(),
             targetPerRefModel_: isStable ? justOne : getNextPriceModel(),
