+++ contracts/fuzz/scenarios/Rebalancing.sol
@@ -299,7 +299,7 @@ contract RebalancingScenario {
         uint8 tokenID,
         uint8 targetNameID,
         uint256 defaultThresholdSeed,
-        uint256 delayUntilDefaultSeed,
+        uint48 delayUntilDefaultSeed,
         bool isColl,
         bool isStable
     ) public onlyDuringState(ScenarioStatus.BEFORE_REBALANCING) {
@@ -322,7 +322,7 @@ contract RebalancingScenario {
         uint8 tokenID,
         uint8 targetNameID,
         uint256 defaultThresholdSeed,
-        uint256 delayUntilDefaultSeed,
+        uint48 delayUntilDefaultSeed,
         bool isColl,
         bool isStable
     ) public onlyDuringState(ScenarioStatus.BEFORE_REBALANCING) {
@@ -917,7 +917,7 @@ contract RebalancingScenario {
         IERC20 erc20,
         bool isStable,
         uint256 defaultThresholdSeed,
-        uint256 delayUntilDefaultSeed,
+        uint48 delayUntilDefaultSeed,
         bytes32 targetName
     ) internal returns (CollateralMock) {
         return
@@ -927,7 +927,7 @@ contract RebalancingScenario {
                 priceTimeout_: 806400,
                 oracleError_: 0.005e18,
                 defaultThreshold_: uint192(between(1, 1e18, defaultThresholdSeed)),
-                delayUntilDefault_: between(1, type(uint48).max / 2, delayUntilDefaultSeed),
+                delayUntilDefault_: uint48(between(1, type(uint48).max / 2, delayUntilDefaultSeed)),
                 targetName_: targetName,
                 refPerTokModel_: isStable ? growing : getNextPriceModel(),
                 targetPerRefModel_: isStable ? justOne : getNextPriceModel(),
